exercises(yacht): change usize to u32

+++ exercises/practice/yacht/.meta/example.zig
@@ -19,8 +19,8 @@ const DiceInt = u3;
 const Dice = [5]DiceInt;
 
 /// Returns the sum of the items in `dice` that equal `n`.
-fn sumOnly(dice: Dice, n: @typeInfo(Category).Enum.tag_type) usize {
-    var result: usize = 0;
+fn sumOnly(dice: Dice, n: @typeInfo(Category).Enum.tag_type) u32 {
+    var result: u32 = 0;
     for (dice) |d| {
         if (d == n) result += n;
     }
@@ -33,13 +33,13 @@ fn sort(dice: Dice) Dice {
     return d;
 }
 
-fn sum(dice: Dice) usize {
-    var result: usize = 0;
+fn sum(dice: Dice) u32 {
+    var result: u32 = 0;
     for (dice) |d| result += d;
     return result;
 }
 
-fn scoreFullHouse(dice: Dice) usize {
+fn scoreFullHouse(dice: Dice) u32 {
     const d = sort(dice);
     const cond =
         (d[0] == d[1] and d[1] == d[2] and d[2] != d[3] and d[3] == d[4]) or
@@ -47,25 +47,25 @@ fn scoreFullHouse(dice: Dice) usize {
     return if (cond) sum(d) else 0;
 }
 
-fn scoreFourOfAKind(dice: Dice) usize {
+fn scoreFourOfAKind(dice: Dice) u32 {
     const d = sort(dice);
     if (d[1] == d[2] and d[2] == d[3]) {
-        if (d[0] == d[1]) return @as(usize, d[0]) * 4;
-        if (d[3] == d[4]) return @as(usize, d[1]) * 4;
+        if (d[0] == d[1]) return @as(u32, d[0]) * 4;
+        if (d[3] == d[4]) return @as(u32, d[1]) * 4;
     }
     return 0;
 }
 
-fn scoreStraight(dice: Dice, straight: Dice) usize {
+fn scoreStraight(dice: Dice, straight: Dice) u32 {
     const d = sort(dice);
     return if (std.mem.eql(DiceInt, &d, &straight)) 30 else 0;
 }
 
-fn scoreYacht(d: Dice) usize {
+fn scoreYacht(d: Dice) u32 {
     return if (d[0] == d[1] and d[1] == d[2] and d[2] == d[3] and d[3] == d[4]) 50 else 0;
 }
 
-pub fn score(dice: Dice, category: Category) usize {
+pub fn score(dice: Dice, category: Category) u32 {
     return switch (category) {
         .ones, .twos, .threes, .fours, .fives, .sixes => sumOnly(dice, @intFromEnum(category) + 1),
         .full_house => scoreFullHouse(dice),
