+++ exercises/practice/yacht/yacht.zig
@@ -13,7 +13,7 @@ pub const Category = enum {
     yacht,
 };
 
-pub fn score(dice: [5]u3, category: Category) usize {
+pub fn score(dice: [5]u3, category: Category) u32 {
     _ = dice;
     _ = category;
     @compileError("please implement the score function");
